requirements.txt
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
filelock==3.19.1
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.45
hydra-core==1.3.2
idna==3.10
Jinja2==3.1.6
joblib==1.5.2
lightning-utilities==0.15.2
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.5
numpy==2.3.3
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
omegaconf==2.3.0
packaging==25.0
pandas==2.3.3
patsy==1.0.1
pillow==11.3.0
platformdirs==4.4.0
protobuf==6.32.1
pydantic==2.11.9
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
scikit-learn==1.7.2
scipy==1.16.2
sentry-sdk==2.39.0
six==1.17.0
smmap==5.0.2
statsmodels==0.14.5
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.8.0
torcheval==0.0.7
torchmetrics==1.8.2
torchvision==0.23.0
tqdm==4.67.1
transforms3d==0.4.2
triton==3.4.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wandb==0.22.1
